Move various utility functions to dedicated submodule
