+++ docs/api/autodoc/spglib.utils.rst
@@ -1,0 +1,8 @@
+spglib.utils module
+===================
+
+.. automodule:: spglib.utils
+   :members:
+   :private-members:
+   :show-inheritance:
+   :undoc-members:
