+++ python/spglib/_internal.py
@@ -11,7 +11,7 @@ from numpy._typing import ArrayLike
 from . import _spglib
 from .error import _set_no_error, _set_or_throw_error
 from .spg import SpgCell, SpglibDataset
-from .spglib import _expand_cell
+from .utils import _expand_cell
 
 __all__ = [
     "get_pointgroup",
